// src/state/actions/types.ts
export const ADD_PRODUCTS = 'ADD_PRODUCTS';

export enum HotelsType {
  GET_ALL_REQUEST = 'GET_ALL_REQUEST',
  GET_ALL_SUCCESS = 'GET_ALL_SUCCESS',
  GET_ALL_FAIL = 'GET_ALL_FAIL',

  ADD_HOTEL_REQUEST = 'ADD_HOTEL_REQUEST',
  ADD_HOTEL_SUCCESS = 'ADD_HOTEL_SUCCESS',
  ADD_HOTEL_FAIL = 'ADD_HOTEL_FAIL'
}